requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.1
attrs==25.3.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
distro==1.9.0
Django==5.2.4
greenlet==3.2.3
griffe==1.7.3
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
hvac==2.3.0
idna==3.10
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
langchain==0.3.26
langchain-core==0.3.71
langchain-text-splitters==0.3.8
langsmith==0.4.8
mcp==1.12.1
openai==1.97.1
openai-agents==0.2.3
orjson==3.11.0
packaging==25.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
requests-toolbelt==1.0.0
rpds-py==0.26.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlparse==0.5.3
sse-starlette==2.4.1
starlette==0.47.2
tenacity==9.1.2
tqdm==4.67.1
types-requests==2.32.4.20250611
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
zstandard==0.23.0
